+++ storage_utils/swift.py
@@ -30,12 +30,14 @@ the `SwiftPath` class.
 from backoff.backoff import with_backoff
 from cached_property import cached_property
 import cStringIO
+import copy
 from functools import wraps
 import json
 import logging
 import os
 import posixpath
 import tempfile
+import threading
 import urlparse
 
 from storage_utils import base
@@ -45,6 +47,7 @@ from storage_utils import utils
 from storage_utils.posix import PosixPath
 from swiftclient import exceptions as swift_exceptions
 from swiftclient import service as swift_service
+from swiftclient import client as swift_client
 from swiftclient.service import SwiftUploadObject
 from swiftclient.utils import generate_temp_url
 
@@ -74,6 +77,10 @@ password = os.environ.get('OS_PASSWORD')
 
 If not set, the ``OS_PASSWORD`` environment variable will be used.
 """
+
+# singleton that collects together auth tokens for storage URLs
+_cached_auth_token_map = {}
+_singleton_lock = threading.Lock()
 
 temp_url_key = os.environ.get('OS_TEMP_URL_KEY')
 """The key for generating temporary URLs
@@ -141,6 +148,34 @@ def update_settings(**settings):
             raise ValueError('invalid setting "%s"' % setting)
         globals()[setting] = value
 
+    _clear_cached_auth_credentials()
+
+
+def _get_or_create_auth_credentials(tenant_name):
+    try:
+        return _cached_auth_token_map[tenant_name]
+    except KeyError:
+        storage_url, auth_token = swift_client.get_auth_keystone(
+            auth_url, username, password,
+            {'tenant_name': tenant_name},
+        )
+        creds = {
+            'os_storage_url': storage_url,
+            'os_auth_token': auth_token
+        }
+
+        # Note: we are intentionally ignoring the rare race condition where
+        # authentication starts in one thread, then settings are updated, and
+        # then authentication finishes in the other.
+        with _singleton_lock:
+            _cached_auth_token_map[tenant_name] = creds
+        return creds
+
+
+def _clear_cached_auth_credentials():
+    with _singleton_lock:
+        _cached_auth_token_map.clear()
+
 
 class SwiftError(Exception):
     """The top-level exception thrown for any swift errors
@@ -173,6 +208,15 @@ class FailedUploadError(UnavailableError):
 
 class UnauthorizedError(SwiftError):
     """Thrown when a 403 response is returned from swift"""
+    pass
+
+
+class AuthenticationError(SwiftError):
+    """Thrown when a client has improper authentication credentials.
+
+    Swiftclient throws this error when trying to authenticate with
+    the keystone client. This is similar to a 401 HTTP response.
+    """
     pass
 
 
@@ -280,10 +324,30 @@ def _propagate_swift_exceptions(func):
                 # Treat this as a FailedUploadError
                 logger.error('upload error in swift put_object operation - %s', str(e))
                 raise FailedUploadError(str(e), e)
+            elif 'Unauthorized.' in str(e):
+                # Swiftclient catches keystone auth errors at
+                # https://github.com/openstack/python-swiftclient/blob/master/swiftclient/client.py#L536 # nopep8
+                # Parse the message since they don't bubble the exception or
+                # provide more information
+                logger.warning('auth error in swift operation - %s', str(e))
+                raise AuthenticationError(str(e), e)
             else:
                 logger.error('unexpected swift error - %s', str(e))
                 raise SwiftError(str(e), e)
 
+    return wrapper
+
+
+def _retry_on_cached_auth_err(func):
+    """Retry a function with cleared auth credentials on AuthenticationError"""
+    @wraps(func)
+    def wrapper(*args, **kwargs):
+        try:
+            return func(*args, **kwargs)
+        except AuthenticationError:
+            logger.info('auth failed, retrying with cleared auth cache')
+            _clear_cached_auth_credentials()
+            return func(*args, **kwargs)
     return wrapper
 
 
@@ -561,6 +625,7 @@ class SwiftPath(base.StorageUtilsPathMixin, str):
         options['os_auth_url'] = auth_url
         options['os_username'] = username
         options['os_password'] = password
+        options.update(**_get_or_create_auth_credentials(self.tenant))
 
         # Merge options with global and local ones
         options = dict(swift_service._default_global_options,
@@ -605,15 +670,23 @@ class SwiftPath(base.StorageUtilsPathMixin, str):
         conn_opts = self._get_swift_connection_options(**options)
         return swift_service.get_conn(conn_opts)
 
+    @_retry_on_cached_auth_err
     @_propagate_swift_exceptions
-    def _swift_connection_call(self, method, *args, **kwargs):
-        """Runs a method call on a ``Connection`` object."""
+    def _swift_connection_call(self, method_name, *args, **kwargs):
+        """Instantiates a ``Connection`` object and runs ``method_name``."""
+        connection = self._get_swift_connection()
+        method = getattr(connection, method_name)
         return method(*args, **kwargs)
 
+    @_retry_on_cached_auth_err
     @_propagate_swift_exceptions
-    def _swift_service_call(self, method, *args, **kwargs):
-        """Runs a method call on a ``SwiftService`` object."""
-        results = method(*args, **kwargs)
+    def _swift_service_call(self, method_name, *args, **kwargs):
+        """Instantiates a ``SwiftService`` object and runs ``method_name``."""
+        method_options = copy.deepcopy(kwargs)
+        service_options = method_options.pop('_service_options', {})
+        service = self._get_swift_service(**service_options)
+        method = getattr(service, method_name)
+        results = method(*args, **method_options)
 
         results = [results] if isinstance(results, dict) else list(results)
         for r in results:
@@ -634,8 +707,7 @@ class SwiftPath(base.StorageUtilsPathMixin, str):
         information about configuring retry logic at the module or method
         level.
         """
-        connection = self._get_swift_connection()
-        headers, content = self._swift_connection_call(connection.get_object,
+        headers, content = self._swift_connection_call('get_object',
                                                        self.container,
                                                        self.resource)
         return content
@@ -759,7 +831,6 @@ class SwiftPath(base.StorageUtilsPathMixin, str):
             ConditionNotMetError: Results were returned, but they did not
                 meet the condition.
         """
-        connection = self._get_swift_connection()
         tenant = self.tenant
         prefix = self.resource
         full_listing = limit is None
@@ -787,11 +858,11 @@ class SwiftPath(base.StorageUtilsPathMixin, str):
             list_kwargs['prefix'] = _with_trailing_slash(list_kwargs['prefix'])
 
         if self.container:
-            results = self._swift_connection_call(connection.get_container,
+            results = self._swift_connection_call('get_container',
                                                   self.container,
                                                   **list_kwargs)
         else:
-            results = self._swift_connection_call(connection.get_account,
+            results = self._swift_connection_call('get_account',
                                                   **list_kwargs)
 
         path_pre = SwiftPath('%s%s' % (self.swift_drive, tenant)) / (self.container or '')
@@ -911,8 +982,7 @@ class SwiftPath(base.StorageUtilsPathMixin, str):
         if not self.resource:
             raise ValueError('can only call download_object on object path')
 
-        service = self._get_swift_service()
-        self._swift_service_call(service.download,
+        self._swift_service_call('download',
                                  container=self.container,
                                  objects=[self.resource],
                                  options={'out_file': out_file})
@@ -997,14 +1067,17 @@ class SwiftPath(base.StorageUtilsPathMixin, str):
                 raise ValueError(
                     '"%s" must be child of download path "%s"' % (obj, self))
 
-        service = self._get_swift_service(object_dd_threads=object_threads,
-                                          container_threads=container_threads)
+        service_options = {
+            'object_dd_threads': object_threads,
+            'container_threads': container_threads
+        }
         download_options = {
             'prefix': _with_trailing_slash(self.resource),
             'out_directory': dest,
             'remove_prefix': True
         }
-        results = self._swift_service_call(service.download,
+        results = self._swift_service_call('download',
+                                           _service_options=service_options,
                                            container=self.container,
                                            objects=objs_to_download.values(),
                                            options=download_options)
@@ -1054,15 +1127,18 @@ class SwiftPath(base.StorageUtilsPathMixin, str):
             raise ValueError('cannot call download on tenant with no container')
         _validate_condition(condition)
 
-        service = self._get_swift_service(object_dd_threads=object_threads,
-                                          container_threads=container_threads)
+        service_options = {
+            'object_dd_threads': object_threads,
+            'container_threads': container_threads
+        }
         download_options = {
             'prefix': _with_trailing_slash(self.resource),
             'out_directory': dest,
             'remove_prefix': True
         }
-        results = self._swift_service_call(service.download,
+        results = self._swift_service_call('download',
                                            self.container,
+                                           _service_options=service_options,
                                            options=download_options)
 
         _check_condition(condition, results)
@@ -1126,8 +1202,6 @@ class SwiftPath(base.StorageUtilsPathMixin, str):
         if not self.container:
             raise ValueError('must specify container when uploading')
 
-        service = self._get_swift_service(object_uu_threads=object_threads,
-                                          segment_threads=segment_threads)
         swift_upload_objects = [
             name for name in to_upload
             if isinstance(name, SwiftUploadObject)
@@ -1145,6 +1219,10 @@ class SwiftPath(base.StorageUtilsPathMixin, str):
             for f in all_files_to_upload
         ])
 
+        service_options = {
+            'object_uu_threads': object_threads,
+            'segment_threads': segment_threads
+        }
         upload_options = {
             'segment_size': segment_size,
             'use_slo': use_slo,
@@ -1152,9 +1230,10 @@ class SwiftPath(base.StorageUtilsPathMixin, str):
             'leave_segments': leave_segments,
             'changed': changed
         }
-        return self._swift_service_call(service.upload,
+        return self._swift_service_call('upload',
                                         self.container,
                                         swift_upload_objects,
+                                        _service_options=service_options,
                                         options=upload_options)
 
     @_swift_retry(exceptions=UnavailableError)
@@ -1174,8 +1253,7 @@ class SwiftPath(base.StorageUtilsPathMixin, str):
             raise ValueError('path must contain a container and resource to '
                              'remove')
 
-        service = self._get_swift_service()
-        return self._swift_service_call(service.delete,
+        return self._swift_service_call('delete',
                                         self.container,
                                         [self.resource])
 
@@ -1199,7 +1277,6 @@ class SwiftPath(base.StorageUtilsPathMixin, str):
         if not self.container:
             raise ValueError('swift path must include container for rmtree')
 
-        service = self._get_swift_service()
         deleting_segments = 'segments' in self.container
         if deleting_segments:
             logger.warning('performing rmtree with segment container "%s". '
@@ -1210,7 +1287,7 @@ class SwiftPath(base.StorageUtilsPathMixin, str):
                            'deleted.', self.container)
 
         if not self.resource:
-            results = self._swift_service_call(service.delete,
+            results = self._swift_service_call('delete',
                                                self.container)
             # Try to delete a segment container since swift client does not
             # do this automatically
@@ -1219,13 +1296,13 @@ class SwiftPath(base.StorageUtilsPathMixin, str):
                                       '.segments_%s' % self.container)
                 for segment_container in segment_containers:
                     try:
-                        self._swift_service_call(service.delete, segment_container)
+                        self._swift_service_call('delete', segment_container)
                     except NotFoundError:
                         pass
             return results
         else:
             to_delete = [p.resource for p in self.list()]
-            return self._swift_service_call(service.delete,
+            return self._swift_service_call('delete',
                                             self.container,
                                             to_delete)
 
@@ -1305,8 +1382,7 @@ class SwiftPath(base.StorageUtilsPathMixin, str):
                 object can't be found.
         """
         stat_objects = [self.resource] if self.resource else None
-        service = self._get_swift_service()
-        result = self._swift_service_call(service.stat,
+        result = self._swift_service_call('stat',
                                           container=self.container,
                                           objects=stat_objects)[0]
 
@@ -1353,8 +1429,7 @@ class SwiftPath(base.StorageUtilsPathMixin, str):
         if not self.container or self.resource:
             raise ValueError('post only works on container paths')
 
-        service = self._get_swift_service()
-        return self._swift_service_call(service.post,
+        return self._swift_service_call('post',
                                         container=self.container,
                                         options=options)